Fix five code-generation bugs and runtime reduce_sum type mismatch
apple_lower.ml:
- out_pids in lower_procedure and result_struct_decls: use dn=0 to find
  return ports (sn=0 finds inputs, not returns)
- declare_outputs for Compound nodes: derive output port set from parent
  eset instead of boundary ins count, so all res_prop ports are
  pre-declared in the outer scope
- res_prop fallback in generic compound: when (sub_gid,0,dp,In) is absent
  from var_map, trace eset to find the producing node
- infer_types pass2: use eset for boundary output ports instead of stale
  boundary outs metadata (often empty)

runtime/sisal_runtime.h:
- sisal_array_reduce_sum: read data as float* (matching how the code
  generator writes array elements) instead of double*, which was
  reinterpreting pairs of floats as one double and producing garbage sums

+++ runtime/sisal_runtime.h
@@ -133,11 +133,11 @@ inline sisal_array_t sisal_array_mul(sisal_array_t a, sisal_array_t b) {
     return res;
 }
 
-inline float sisal_array_reduce_sum(sisal_array_t a) { 
-    double s = 0;
-    double* d = (double*)a.data;
+inline float sisal_array_reduce_sum(sisal_array_t a) {
+    float s = 0;
+    float* d = (float*)a.data;
     for(uint64_t i=0; i<a.size; i++) s += d[i];
-    return (float)s;
+    return s;
 }
 
 inline int32_t sisal_array_reduce_int_product(sisal_array_t a) {
